Fix dead letter queue (#83)
HPT-567 Determine why there are dead letters appearing in the Activity history queue

The dead letters were appearing because there were a number of activity history events that this listener has not been handling. These event types are:

    ContactDetailEditedEvent
    NoteCreatedAgainstAssetEvent
    NoteCreatedAgainstTenureEvent
    NoteCreatedAgainstPersonEvent

Currently when these events are received the listener throws an ArgumentException because it cannot match them to any specified events in the EventFactory.cs - this is a lambda failure which ultimately leads the events to the dead letter queue.

To keep things working as they are now but also clear the dead letter queue, this PR makes these events log a Will not process message of type: warning message and not take further action. It's assumed that there's not a pressing business need to properly handle these events as they have been failing for months.