Address Dockerfile Shell vs Exec form warning.
Shell instead of the exec form for CMD & ENTRYPOINT within Dockerfiles can lead to executable runs as a child process to a shell, which does not pass OS signals. This can cause problems when trying to gracefully stop containers because the main process will not receive the signal intended to terminate it.